+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed: `selector-no-type` does a better job when ignoring descendant and compound selectors.
 - Fixed: `block-closing-brace-newline-before` no longer delivers false positives for extra semicolon.
 - Fixed: `font-family-name-quotes` is now case insensitive when hunting for font-family properties.
+- Fixed: `number-zero-length-no-unit` now ignores `deg`, `grad`, `turn` and `rad` units.
 
 # 6.3.2